adding principal to the url
